feat: Add support for GFF file in IVAR_VARIANTS process
This commit modifies the IVAR_VARIANTS process in the nf-core/ivar/variants module to add support for a GFF file. The `gff_file` parameter is introduced, allowing users to specify a GFF file for variant calling. If no GFF file is provided, the process will run without it. This enhancement provides more flexibility in variant calling analysis.