fix: Update paths for md5.sh and reference.fasta files
This commit updates the paths for the `md5.sh` script and `reference.fasta` file in the Makefile, test.config, and nextflow.config files. The paths are changed to use the `assets` directory instead of the `test` directory. This change ensures that the correct files are referenced and improves the organization of the project structure.